backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
cryptography==41.0.4
psycopg2-binary==2.9.9
